ref(replay): consolidate replay-duration and replay-id into shared utilities
- Delete src/lib/replay-duration.ts: move formatDurationCompact,
  formatDurationVerbose, and formatDurationCompactMs to
  src/lib/formatters/time-utils.ts where other duration helpers live.
- Delete src/lib/replay-id.ts: add tryNormalizeHexId to hex-id.ts
  (non-throwing variant), move getReplayIdFromEvent and collectReplayIds
  to replay-search.ts (replay domain logic).
- Replace local formatRelativeOffset in replay/view.ts with the shared
  formatDurationCompactMs from time-utils.ts.
- Fix event/view.ts replayHint reference to undefined 'event' variable
  introduced during merge conflict resolution (use fetchedEvents[0]).
