+++ plugins/sentry-cli/skills/sentry-cli/references/replay.md
@@ -1,6 +1,6 @@
 ---
 name: sentry-cli-replay
-version: 0.31.0-dev.0
+version: 0.32.0-dev.0
 description: Search and inspect Session Replays
 requires:
   bins: ["sentry"]
